#631, #632: Unit test updated. Sonar fix.

+++ lionengine-core-awt/src/main/java/com/b3dgs/lionengine/awt/graphic/GraphicAwt.java
@@ -138,8 +138,8 @@ final class GraphicAwt implements Graphic
         }
         else
         {
-            final double ax = dx2 - sx1 + angleX;
-            final double ay = dy2 - sy1 + angleY;
+            final double ax = dx2 - sx1 + (double) angleX;
+            final double ay = dy2 - sy1 + (double) angleY;
 
             final AffineTransform old = g.getTransform();
             final AffineTransform tx = AffineTransform.getRotateInstance(Math.toRadians(angle), ax, ay);
